pyproject.toml
[project]
name = "github-copilot-air-pollution-app"
version = "0.1.0"
description = "Air pollution data viewer using OpenWeatherMap APIs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
]
